Add files via upload
moved the substitution of wpa_supplicant and dhcpcd for NetworkManager to the end, thus saving one step where the user would have to manually interact with the script mid-execution to use nmtui to connect to the internet again. failure to do so would cause the script to not work. added a few more quality of life packages.